+++ examples/demo_agents/demo_SAC.py
@@ -15,7 +15,7 @@ from rlberry.envs.basewrapper import Wrapper
 # from rlberry.envs import gym_make
 from rlberry.manager import plot_writer_data, AgentManager
 from rlberry.envs.benchmarks.ball_exploration import PBall2D
-from rlberry.agents.torch import SACAgent
+from rlberry.agents.experimental.torch import SACAgent
 import gym